Fixed defect https://github.com/upgrad-edu/Course_4_Assignment/issues/3

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -92,8 +92,19 @@ public class ImageController {
     //The method first needs to convert the list of all the tags to a string containing all the tags separated by a comma and then add this string in a Model type object
     //This string is then displayed by 'edit.html' file as previous tags of an image
     @RequestMapping(value = "/editImage")
-    public String editImage(@RequestParam("imageId") Integer imageId, Model model) {
+    public String editImage(@RequestParam("imageId") Integer imageId, Model model, HttpSession session) {
         Image image = imageService.getImage(imageId);
+
+        User user = (User) session.getAttribute( "loggeduser" );
+
+        if (user.getId() != image.getUser().getId()) {
+            String error = "Only the owner of the image can edit the image";
+            model.addAttribute("image", image);
+            model.addAttribute("editError", error);
+            model.addAttribute("tags", image.getTags());
+
+            return "images/image";
+        }
 
         String tags = convertTagsToString(image.getTags());
         model.addAttribute("image", image);
@@ -140,7 +151,20 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, Model model, HttpSession session) {
+
+        Image image = imageService.getImage( imageId );
+
+        User user = (User) session.getAttribute( "loggeduser" );
+
+        if (user.getId() != image.getUser().getId()) {
+            String error = "Only the owner of the image can delete the image";
+            model.addAttribute("image", image);
+            model.addAttribute("deleteError", error);
+            model.addAttribute("tags", image.getTags());
+
+            return "images/image";
+        }
         imageService.deleteImage(imageId);
         return "redirect:/images";
     }
